Switched from static const to constexpr.

+++ io/emulation/gamepads/classic-gamepad.h
@@ -23,7 +23,7 @@ private:
 	void syncSticks();
 
 private:
-	std::array<int16_t, 6> m_axis;
+	std::array<int32_t, 6> m_axis;
 };
 }
 }
